requirements.txt
python-dotenv>=1.0.0
mcp>=1.0.0
